requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.2
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
extensions==0.4
gunicorn==20.1.0
idna==2.10
Markdown==3.3.4
parse-torrent-title==2.3
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
